First stage finish

+++ src/Facade/AdminFacade.java
@@ -3,18 +3,21 @@ package Facade;
 import java.util.ArrayList;
 
 import Exeptions.FacadeException;
+import Exeptions.LoginManagerException;
+import Facade.LoginManager.ClientType;
 import beans.Company;
 import beans.Coupon;
 import beans.Customer;
 
 public class AdminFacade extends ClientFacade {
-    private static CompanyFacade companyFacade;
-    private static CustomerFacade customerFacade;
+    // private static CompanyFacade companyFacade;
+    // private static CustomerFacade customerFacade;
+    private static LoginManager loginManager = LoginManager.getInstance();
 
     public AdminFacade() {
         super();
-        companyFacade = new CompanyFacade();
-        customerFacade = new CustomerFacade();
+        // companyFacade = new CompanyFacade();
+        // customerFacade = new CustomerFacade();
     }
 
     @Override
@@ -32,11 +35,18 @@ public class AdminFacade extends ClientFacade {
             throw new FacadeException("Email = " + company.getEmail() + " is in use with another user, try another");
         if (companiesDAO.isCompanyNameExist(company.getName())) {
             throw new FacadeException("Name = " + company.getName() + " is in use with another user, try another");
-
         }
         companiesDAO.addCompany(company);
-        for (Coupon coupon : company.getCoupons())
-            companyFacade.addCoupon(coupon);
+        CompanyFacade companyFacade;
+        try {
+            companyFacade = (CompanyFacade) loginManager.login(company.getEmail(), company.getPassword(),
+                    ClientType.COMPANY);
+            for (Coupon coupon : company.getCoupons()) {
+                companyFacade.addCoupon(coupon);
+            }
+        } catch (LoginManagerException e) {
+            System.out.println(e.getMessage());
+        }
     }
 
     public void updateCompany(Company company) throws FacadeException {
@@ -46,9 +56,36 @@ public class AdminFacade extends ClientFacade {
         Company oldCompany = companiesDAO.getCompany(company.getId());
         if (!oldCompany.getName().equals(company.getName())) {
             throw new FacadeException("Company id = " + company.getId() + " can't change name");
-
         }
+        ArrayList<Coupon> oldCoupons = couponsDAO.getAllCompanyCoupons(company.getId());
+        ArrayList<Coupon> newCoupons = company.getCoupons();
         companiesDAO.updateCompany(company);
+        // Delete all old coupons and add all new coupons
+        for (Coupon coupon : newCoupons) {
+            if (!oldCoupons.contains(coupon)) {
+                try {
+                    CompanyFacade companyFacade;
+                    companyFacade = (CompanyFacade) loginManager.login(company.getEmail(), company.getPassword(),
+                            ClientType.COMPANY);
+                    companyFacade.addCoupon(coupon);
+                } catch (LoginManagerException e) {
+                    System.out.println(e.getMessage());
+                }
+            }
+        }
+        for (Coupon coupon : oldCoupons) {
+            if (!newCoupons.contains(coupon)) {
+                CompanyFacade companyFacade;
+                try {
+                    companyFacade = (CompanyFacade) loginManager.login(company.getEmail(), company.getPassword(),
+                            ClientType.COMPANY);
+                    companyFacade.deleteCoupon(coupon.getId());
+                } catch (LoginManagerException e) {
+                    System.out.println(e.getMessage());
+                }
+            }
+        }
+
     }
 
     public void deleteCompany(int companyID) throws FacadeException {
@@ -71,10 +108,9 @@ public class AdminFacade extends ClientFacade {
         return companies;
     }
 
-    public Company getOneCompany(int companyID) {
+    public Company getOneCompany(int companyID) throws FacadeException {
         if (!companiesDAO.isCompanyExist(companyID)) {
-            // TODO exception
-            return null;
+            throw new FacadeException("Company id = " + companyID + " doesn't exist");
         }
         Company company = companiesDAO.getCompany(companyID);
         company.setCoupons(couponsDAO.getAllCompanyCoupons(companyID));
@@ -83,48 +119,65 @@ public class AdminFacade extends ClientFacade {
 
     public void addCustomer(Customer customer) throws FacadeException {
         if (customersDAO.isCustomerExist(customer.getId()))
-            throw new FacadeException("Customer already exist(by ID)");
+            throw new FacadeException("Customer id = " + customer.getId() + " already exist");
         if (customersDAO.isCustomerEmailExist(customer.getEmail())) {
-            // TODO throw new Exception here
-            System.out.println("Customer id = " + customer.getId() + " already exist");
-            return;
+            throw new FacadeException("Customer Email = " + customer.getEmail() + " already exist");
         }
         customersDAO.addCustomer(customer);
-        ArrayList<Coupon> coupons = couponsDAO.getAllCustomerCoupons(customer.getId());
-        for (Coupon coupon : customer.getCoupons()) {
-            if (!coupons.contains(coupon))// this is better than couponFacade.purchaseCoupon -> don't waste connections
-                couponsDAO.addCouponPurchase(customer.getId(), coupon.getId());
+        CustomerFacade customerFacade;
+        try {
+            customerFacade = (CustomerFacade) loginManager.login(customer.getEmail(), customer.getPassword(),
+                    ClientType.CUSTOMER);
+            for (Coupon coupon : customer.getCoupons())
+                customerFacade.purchaseCoupon(coupon);
+        } catch (LoginManagerException | FacadeException e) {
+            System.out.println(e.getMessage());
         }
-        System.out.println("Customer id = " + customer.getId() + " inserted ok.");
+
     }
 
-    public void updateCustomer(Customer customer) {
+    public void updateCustomer(Customer customer) throws FacadeException {
         if (customer == null) {
             System.out.println("argument was null->bug");
         }
         if (!customersDAO.isCustomerExist(customer.getId())) {
-            // TODO throw new exception here
-            System.out.println("Customer id = " + customer.getId() + " doesn't exist");
-            return;
+            throw new FacadeException("Customer id = " + customer.getId() + " doesn't exist");
         }
         ArrayList<Coupon> oldCoupons = couponsDAO.getAllCustomerCoupons(customer.getId());
         ArrayList<Coupon> newCoupons = customer.getCoupons();
         customersDAO.updateCustomer(customer);
-        // Delete all old purchased and add all new purchases
-        for (Coupon coupon : newCoupons) {
-            if (!oldCoupons.contains(coupon))
-                couponsDAO.addCouponPurchase(customer.getId(), coupon.getId());
-        }
-        for (Coupon coupon : oldCoupons) {
-            if (!newCoupons.contains(coupon))
-                couponsDAO.deleteCouponPurchase(customer.getId(), coupon.getId());
+        CustomerFacade customerFacade;
+        try {
+            customerFacade = (CustomerFacade) loginManager.login(customer.getEmail(), customer.getPassword(),
+                    ClientType.CUSTOMER);
+            // Delete all old purchased and add all new purchases
+            for (Coupon coupon : newCoupons) {
+                if (!oldCoupons.contains(coupon)) {
+                    customerFacade.purchaseCoupon(coupon);
+                }
+            }
+            for (Coupon coupon : oldCoupons) {
+                if (!newCoupons.contains(coupon)) {
+                    couponsDAO.deleteCouponPurchase(customer.getId(), coupon.getId());
+                    coupon.setAmount(coupon.getAmount() + 1);
+                    couponsDAO.updateCoupon(coupon);
+                }
 
+            }
+        } catch (LoginManagerException e) {
+            System.out.println(e.getMessage());
         }
     }
 
-    public void deleteCustomer(int customerID) {
+    public void deleteCustomer(int customerID) throws FacadeException {
+        if (!customersDAO.isCustomerExist(customerID)) {
+            throw new FacadeException("Customer id = " + customerID + " doesn't exist");
+        }
         customersDAO.deleteCustomer(customerID);
-        couponsDAO.deleteAllCustomerCouponsPurchases(customerID);
+        if (couponsDAO.getAllCustomerCoupons(customerID).size() == 0)
+            ;
+        else
+            couponsDAO.deleteAllCustomerCouponsPurchases(customerID);
     }
 
     public ArrayList<Customer> getAllCustomers() {
@@ -135,10 +188,9 @@ public class AdminFacade extends ClientFacade {
         return customers;
     }
 
-    public Customer getOneCustomer(int customerID) {
+    public Customer getOneCustomer(int customerID) throws FacadeException {
         if (!customersDAO.isCustomerExist(customerID)) {
-            // TODO exception
-            return null;
+            throw new FacadeException("Customer id = " + customerID + " doesn't exist");
         }
         Customer customer = customersDAO.getOneCustomer(customerID);
         customer.setCoupons(couponsDAO.getAllCustomerCoupons(customerID));
